# api-snapshots/openapi.json
{
  "components": {
    "schemas": {
      "AttachmentOut": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "media_type": {
            "title": "Media Type",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "reference": {
            "title": "Reference",
            "type": "string"
          },
          "size": {
            "title": "Size",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "name",
          "media_type",
          "size",
          "reference"
        ],
        "title": "AttachmentOut",
        "type": "object"
      },
      "ErrorBody": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "additionalProperties": true,
            "title": "Detail",
            "type": "object"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorEnvelope": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
      "ExecutionBackend": {
        "enum": [
          "host",
          "docker"
        ],
        "title": "ExecutionBackend",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "InteractionModeOut": {
        "properties": {
          "applied": {
            "title": "Applied",
            "type": "string"
          },
          "mode": {
            "title": "Mode",
            "type": "string"
          }
        },
        "required": [
          "mode",
          "applied"
        ],
        "title": "InteractionModeOut",
        "type": "object"
      },
...
      "ModelSource": {
        "enum": [
          "gateway",
          "native"
        ],
        "title": "ModelSource",
        "type": "string"
      },
...
      "PrivacyMode": {
        "enum": [
          "none",
          "surrogate"
        ],
        "title": "PrivacyMode",
        "type": "string"
      },
...
      "RenameIn": {
        "properties": {
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title"
        ],
        "title": "RenameIn",
        "type": "object"
      },
...
      "SessionOut": {
        "properties": {
          "activity": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Activity"
          },
          "created_at": {
            "title": "Created At",
            "type": "integer"
          },
          "execution_backend": {
            "$ref": "#/components/schemas/ExecutionBackend",
            "default": "host"
          },
          "harness": {
            "title": "Harness",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "interaction_mode": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/InteractionModeOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "last_activity_at": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Activity At"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "model_source": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ModelSource"
              },
              {
                "type": "null"
              }
            ]
          },
          "native_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Native Id"
          },
          "policy_revision": {
            "default": 1,
            "title": "Policy Revision",
            "type": "integer"
          },
          "privacy_mode": {
            "$ref": "#/components/schemas/PrivacyMode",
            "default": "none"
          },
          "project_path": {
            "title": "Project Path",
            "type": "string"
          },
          "reasoning_effort": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reasoning Effort"
          },
          "state": {
            "title": "State",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "integer"
          },
          "worktree": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Worktree"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "id",
          "harness",
          "native_id",
          "title",
          "project_path",
          "created_at",
          "updated_at",
          "state",
          "model"
        ],
        "title": "SessionOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "Worktree": {
        "properties": {
          "base_commit": {
            "title": "Base Commit",
            "type": "string"
          },
          "base_ref": {
            "title": "Base Ref",
            "type": "string"
          },
          "discard": {
            "default": false,
            "title": "Discard",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "integrated_commit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Integrated Commit"
          },
          "integrated_head": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Integrated Head"
          },
          "integrated_index": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Integrated Index"
          },
          "integrated_target": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Integrated Target"
          },
          "integrated_tree": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Integrated Tree"
          },
          "path": {
            "title": "Path",
            "type": "string"
          },
          "pending_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pending Id"
          },
          "pending_integration": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pending Integration"
          },
          "relative_path": {
            "default": ".",
            "title": "Relative Path",
            "type": "string"
          },
          "repository": {
            "title": "Repository",
            "type": "string"
          },
          "source_path": {
            "title": "Source Path",
            "type": "string"
          },
          "state": {
            "default": "preparing",
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "id",
          "path",
          "repository",
          "source_path",
          "base_ref",
          "base_commit"
        ],
        "title": "Worktree",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Mandri Daemon",
    "version": "0.1.0b0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/sessions": {
      "get": {
        "operationId": "list_sessions",
        "parameters": [
          {
            "in": "query",
            "name": "harness",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Harness"
            }
          },
          {
            "in": "query",
            "name": "state",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "State"
            }
          },
          {
            "in": "query",
            "name": "project_path",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Project Path"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SessionOut"
                  },
                  "title": "Response List Sessions",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Sessions",
        "tags": [
          "sessions"
        ]
      }
    },
    "/v1/sessions/{session_id}": {
      "delete": {
        "operationId": "delete_session",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "description": "Session identifier (36-char lowercase uuid4)",
              "pattern": "^[0-9a-f][0-9a-f-]{34}[0-9a-f]$",
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "purge",
            "required": false,
            "schema": {
              "default": false,
              "title": "Purge",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "discard_worktree",
            "required": false,
            "schema": {
              "default": false,
              "title": "Discard Worktree",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Resource not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Session",
        "tags": [
          "sessions"
        ]
      },
      "get": {
        "operationId": "get_session",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "description": "Session identifier (36-char lowercase uuid4)",
              "pattern": "^[0-9a-f][0-9a-f-]{34}[0-9a-f]$",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Resource not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Session",
        "tags": [
          "sessions"
        ]
      },
      "patch": {
        "operationId": "rename_session",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "description": "Session identifier (36-char lowercase uuid4)",
              "pattern": "^[0-9a-f][0-9a-f-]{34}[0-9a-f]$",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RenameIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Resource not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Rename Session",
        "tags": [
          "sessions"
        ]
      }
    },
    "/v1/sessions/{session_id}/attachments": {
      "post": {
        "operationId": "upload_attachment",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "description": "Session identifier (36-char lowercase uuid4)",
              "pattern": "^[0-9a-f][0-9a-f-]{34}[0-9a-f]$",
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "name",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 1,
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/octet-stream": {
              "schema": {
                "format": "binary",
                "type": "string"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttachmentOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Attachment",
        "tags": [
          "attachments"
        ]
      }
    },
...
    "/v1/sessions/{session_id}/files": {
      "get": {
        "operationId": "download_session_file",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "description": "Session identifier (36-char lowercase uuid4)",
              "pattern": "^[0-9a-f][0-9a-f-]{34}[0-9a-f]$",
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "path",
            "required": true,
            "schema": {
              "maxLength": 4096,
              "minLength": 1,
              "title": "Path",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Download Session File",
        "tags": [
          "attachments"
        ]
      }
    },
...
  }
}